Fix WiFi settings in settings server
- Move WiFi functionality into dedicated module (gonotego.settings.wifi)
- Update system_commands.py to use the new WiFi module
- Update server.py to properly import and use the WiFi module
- Fix WiFi settings handling in server.py to use the proper modules
